Fix default constructors in calculator not casting to type

+++ Ejercicio 1/main.cpp
@@ -2,5 +2,5 @@
 
 void main()
 {
-	FractionCalculator<int> calculator = FractionCalculator<int>();
+	FractionCalculator<short> calculator = FractionCalculator<short>();
 }
